opt/optbuilder: refactor builder.go
builder.go was getting too large (well over 1000 lines of code).
This commit breaks builder.go up into logical components:

distinct.go
groupby.go (already refactored in #22917)
join.go
project.go
scalar.go
select.go

No code has been modified - just moved around.

Release note: None
